pyproject.toml
[project]
name = "mcp-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.1",
    "mcp>=1.26.0",
    "openai-agents>=0.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
